Fix: Use persistent path for fulfillment history file
This commit fixes a critical bug that caused errors when running the application from a read-only network share. The application was previously attempting to read and write the `fulfillment_history.csv` file to the current working directory, which would fail if the application was located on a file server where the user lacks write permissions.

Key Changes:
- **Centralized Path Utilities:** The `get_persistent_data_path` and `resource_path` helper functions have been moved from `gui_main.py` to a new shared module, `shopify_tool/utils.py`, to improve code organization and prevent potential circular dependencies.
- **Persistent History Path:** The logic in `shopify_tool/core.py` has been updated to use `get_persistent_data_path` for the fulfillment history file. This ensures that `fulfillment_history.csv` is always stored in the user's local application data directory, regardless of where the application itself is launched from.

This change should resolve the `EmptyDataError` and make the application's file handling more robust and predictable, especially in enterprise environments.
