+++ gui_main.py
@@ -50,34 +50,9 @@ class ToolTip:
             self.tooltip_window.destroy()
         self.tooltip_window = None
 
-def get_persistent_data_path(filename):
-    """
-    Generates a path to a file in a persistent application data folder.
-    Creates the folder if it doesn't exist.
-    """
-    # Use APPDATA for Windows, or user's home directory for other platforms
-    app_data_path = os.getenv('APPDATA') or os.path.expanduser("~")
-    app_dir = os.path.join(app_data_path, "ShopifyFulfillmentTool")
-
-    try:
-        os.makedirs(app_dir, exist_ok=True)
-    except OSError as e:
-        # Fallback to current directory if AppData is not writable
-        logger.error(f"Could not create AppData directory: {e}. Falling back to local directory.")
-        app_dir = "."
-
-    return os.path.join(app_dir, filename)
-
-def resource_path(relative_path):
-    """ Get absolute path to resource, works for dev and for PyInstaller """
-    try:
-        base_path = sys._MEIPASS
-    except Exception:
-        base_path = os.path.abspath(".")
-    return os.path.join(base_path, relative_path)
-
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '.')))
 from shopify_tool import core
+from shopify_tool.utils import get_persistent_data_path, resource_path
 from shopify_tool.analysis import recalculate_statistics, toggle_order_fulfillment
 from gui.report_builder_window import ReportBuilderWindow
 from gui.settings_window import SettingsWindow
